Renames the 'requiresAuth' folder to 'auth'
