+++ client-side-auth-iii/src/auth/requiresAuth.js
